+++ Editor/Idol/ToonPBRSetupCheck.cs
@@ -513,6 +513,8 @@ namespace ToonNPR.EditorTools
             {
                 ("_BumpMap",        "_NormalMapOn",          "Normal Map"),
                 ("_NPRMap",         "_NPRMapOn",             "NPR Map"),
+                ("_FabricMap",      "_FabricMapOn",          "Fabric Map"),
+                ("_AnisotropyMap",  "_AnisotropyMapOn",      "Anisotropy Map"),
                 ("_BentNormalMap",  "_BentNormalOn",         "Bent Normal"),
                 ("_ShadeNormalMap", "_ShadeNormalStrength",  "Shade Normal"),
                 ("_CurvatureMap",   "_CurvatureSoftness",    "Curvature Map"),